refactor(calendar): move month matrix helpers out of day-rail
